requirements.txt
absl-py==1.4.0
aiohttp @ file:///croot/aiohttp_1694181079612/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
array-record==0.4.0
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1691763562544/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1696764509355/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1633010882559/work
certifi==2023.7.22
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1696431134479/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
contourpy==1.1.1
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687014355/work
cycler==0.12.1
dm-tree==0.1.8
etils==1.3.0
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1685164110706/work
fonttools==4.43.1
frozenlist @ file:///croot/frozenlist_1670004507010/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1629296548061/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1675825735130/work
google-pasta==0.2.0
googleapis-common-protos==1.61.0
grpcio @ file:///croot/grpc-suite_1681912592597/work
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1604753633781/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1688754491823/work
importlib-resources==6.1.0
jax==0.4.13
keras @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/keras_1682445665871/work/keras-2.12.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver==1.4.5
libclang==16.0.6
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1696627721496/work
matplotlib==3.7.3
ml-dtypes==0.2.0
multidict @ file:///croot/multidict_1665674239670/work
nptyping==2.5.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1651020413938/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
opencv-python==4.8.1.78
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1696448916724/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.0.3
Pillow==10.1.0
promise==2.3
protobuf==3.20.3
psutil==5.9.6
pyasn1 @ file:///home/conda/feedstock_root/build_artifacts/pyasn1_1694615621498/work
pyasn1-modules @ file:///home/conda/feedstock_root/build_artifacts/pyasn1-modules_1695107857548/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1689721553971/work
pynrrd==1.0.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==3.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2023.3.post1
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1653073867187/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorboard_1682445826165/work/tensorboard-2.12.1-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-base_1682700874376/work/tensorflow_pkg/tensorflow-2.12.0-cp38-cp38-linux_x86_64.whl
tensorflow-datasets==4.9.2
tensorflow-estimator @ file:///home/builder/mesters/opt/envs/tensorflow/conda-bld/tensorflow-estimator_1682445976941/work/tensorflow_estimator-2.12.0-py2.py3-none-any.whl
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1682317048417/work
toml==0.10.2
tqdm==4.66.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata==2023.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1696434355823/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1610094884173/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1648966513179/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
